+++ lib_src/src/UDPConnection.cpp
@@ -31,7 +31,7 @@ bool Network::UDPConnection::sendPacket(IPacket const &packet)
     boost::mutex::scoped_lock   lock{_ioMutex};
     PacketBuffer	            _finalBuffer{};
     PacketBuffer                toSend{packet.getData()};
-    PacketSize                  packetSize{toSend.size()};
+    PacketSize                  packetSize{static_cast<int>(toSend.size())};
 
     if (_toSendBuffer.reserve() < packetSize)
     {
